+++ lib/LocalizedTerritory.php
@@ -18,6 +18,6 @@ class LocalizedTerritory extends LocalizedObject
 	protected function get_name(): string
 	{
 		/** @phpstan-ignore-next-line */
-		return $this->locale['territories'][$this->target->code];
+		return $this->locale['territories'][$this->target->code->value];
 	}
 }