+++ tests/RepositoryTest.php
@@ -5,6 +5,8 @@ namespace Test\ICanBoogie\CLDR;
 use ICanBoogie\CLDR\CurrencyFormatter;
 use ICanBoogie\CLDR\ListFormatter;
 use ICanBoogie\CLDR\Locale\ListPattern;
+use ICanBoogie\CLDR\LocaleId;
+use ICanBoogie\CLDR\LocaleNotAvailable;
 use ICanBoogie\CLDR\NumberFormatter;
 use ICanBoogie\CLDR\Plurals;
 use ICanBoogie\CLDR\Provider;
@@ -84,24 +86,45 @@ final class RepositoryTest extends TestCase
 		$this->assertSame("one, two, and three", $this->sut->format_list($list, $list_pattern));
 	}
 
+	public function test_locale_for_using_string(): void
+	{
+		$actual = $this->sut->locale_for('fr-BE');
+
+		$this->assertEquals('fr-BE', $actual->id->value);
+	}
+
+	public function test_locale_for_using_id(): void
+	{
+		$actual = $this->sut->locale_for(LocaleId::of('fr-BE'));
+
+		$this->assertEquals('fr-BE', $actual->id->value);
+	}
+
+	public function test_locale_for_fails_on_unavailable_id(): void
+	{
+		$this->expectException(LocaleNotAvailable::class);
+
+		$this->sut->locale_for('foo');
+	}
+
 	public function test_territory_for_using_string(): void
 	{
 		$actual = $this->sut->territory_for('CA');
 
-		$this->assertEquals('CA', $actual->code);
+		$this->assertEquals('CA', $actual->code->value);
 	}
 
 	public function test_territory_for_using_code(): void
 	{
 		$actual = $this->sut->territory_for(TerritoryCode::of('CA'));
 
-		$this->assertEquals('CA', $actual->code);
+		$this->assertEquals('CA', $actual->code->value);
 	}
 
 	public function test_territory_for_fails_on_undefined_code(): void
 	{
 		$this->expectException(TerritoryNotDefined::class);
 
-		$this->sut->territory_for('ZZZ');
+		$this->sut->territory_for('foo');
 	}
 }